# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoCopyAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz answers and get results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
  /quiz/results/{id}:
    get:
      operationId: getQuizResult
      tags: [quiz]
      summary: Get a specific quiz result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── AI Generation Tools ───────────────────────────────────────────────────────
...
  /brand-profiles:
    get:
      operationId: listBrandProfiles
      tags: [brand-profiles]
      summary: List brand profiles
      responses:
        "200":
          description: List of brand profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandProfile"
    post:
      operationId: createBrandProfile
      tags: [brand-profiles]
      summary: Create brand profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
  /brand-profiles/{id}:
    get:
      operationId: getBrandProfile
      tags: [brand-profiles]
      summary: Get brand profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
        "404":
          description: Not found
    put:
      operationId: updateBrandProfile
      tags: [brand-profiles]
      summary: Update brand profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
    delete:
      operationId: deleteBrandProfile
      tags: [brand-profiles]
      summary: Delete brand profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Video Projects ───────────────────────────────────────────────────────────
...
  /calendar:
    get:
      operationId: listCalendarItems
      tags: [calendar]
      summary: List calendar items
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: YYYY-MM
      responses:
        "200":
          description: List of calendar items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarItem"
    post:
      operationId: createCalendarItem
      tags: [calendar]
      summary: Create calendar item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarItem"
  /calendar/{id}:
    put:
      operationId: updateCalendarItem
      tags: [calendar]
      summary: Update calendar item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarItem"
    delete:
      operationId: deleteCalendarItem
      tags: [calendar]
      summary: Delete calendar item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /calendar/ai-fill:
    post:
      operationId: aiFillCalendar
      tags: [calendar]
      summary: AI-fill empty calendar slots for a month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiFillCalendarRequest"
      responses:
        "200":
          description: Generated and saved items
          content:
            application/json:
              schema:
                type: object
                required: [items, creditsUsed]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CalendarItem"
                  creditsUsed:
                    type: number

  # ── Generation Feedback ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
    QuizResult:
      type: object
      required: [id, overallScore, stage, categoryScores, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        overallScore:
          type: number
        stage:
          type: string
        categoryScores:
          type: object
          additionalProperties:
            type: number
        generatedPlan:
          type: ["string", "null"]
        createdAt:
          type: string
...
    BrandProfile:
      type: object
      required: [id, userId, name, tone, platforms, isDefault, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        industry:
          type: string
          nullable: true
        productDescription:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        tone:
          type: string
        platforms:
          type: array
          items:
            type: string
        brandValues:
          type: string
          nullable: true
        competitors:
          type: string
          nullable: true
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BrandProfileInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: string
          nullable: true
        productDescription:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        tone:
          type: string
          nullable: true
        platforms:
          type: array
          items:
            type: string
          nullable: true
        brandValues:
          type: string
          nullable: true
        competitors:
          type: string
          nullable: true
        isDefault:
          type: boolean
          nullable: true

    # ── Video Projects ──────────────────────────────────────────────────────────
...
    CalendarItem:
      type: object
      required: [id, userId, scheduledDate, platform, contentType, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandProfileId:
          type: integer
          nullable: true
        scheduledDate:
          type: string
        platform:
          type: string
        contentType:
          type: string
        title:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, scheduled, published]
        generationId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    CalendarItemInput:
      type: object
      required: [scheduledDate, platform, contentType]
      properties:
        brandProfileId:
          type: integer
          nullable: true
        scheduledDate:
          type: string
        platform:
          type: string
        contentType:
          type: string
        title:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, scheduled, published]
          nullable: true
    AiFillCalendarRequest:
      type: object
      required: [month, platforms, postsPerWeek]
      properties:
        month:
          type: string
          description: YYYY-MM
        platforms:
          type: array
          items:
            type: string
        postsPerWeek:
          type: integer
          minimum: 1
          maximum: 21
        brandProfileId:
          type: integer
          nullable: true
        topic:
          type: string
          nullable: true

    # ── Feedback ────────────────────────────────────────────────────────────────
...